+++ cats/__init__.py
@@ -4,4 +4,4 @@ from .plugins import *
 from .types import *
 from .utils import *
 
-version = '4.2.0'
+version = '4.2.1'